apps/bot/requirements.txt
discord.py
datetime
upstash_redis
python-dotenv
pytz